+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+1.5.5 (unreleased)
+++++++++++++++++++
+
+* Fixed bug with ``fill()`` causing ``onchange`` event to trigger twice.
+* Fixed headless mode for latest Selenium with Firefox/Chrome
+
 1.5.4 (2023-04-19)
 ++++++++++++++++++